requirements.txt
scapy==2.6.1
numpy==2.2.3
scikit-learn==1.6.1
redis==5.2.1
fastapi==0.115.8
uvicorn==0.34.0
psycopg2-binary
python-dotenv==1.0.1
joblib==1.4.2
ruff==0.9.7
pytest==8.3.4
